requirements.txt
ffmpeg-python
python-vlc